# docs/rest_v1_openapi.yaml
openapi: 3.0.2
servers:
  - url: http://localhost:3370
    description: Local Linstor controller
security:
  - bearerAuth: []
  - {}
info:
  description: |
    Linstor REST API V1

    The V1 rest api of Linstor should stay compatible and only additions are made to the API,
    If there are breaking changes or redesigned a new major REST API version will be issued.

    Server runs per default on port `3370` on `::` ipv6 and ipv4.

    To change the bind address or port you can use the following linstor client commands:
    ```
    linstor controller set-property REST/bindAddress 127.0.0.1
    linstor controller set-property REST/port 8080
    ```

    After setting this properties restart the controller and the new values should be used.
  version: 1.29.0
  title: Linstor REST API
  contact:
    email: rene.peinthor@linbit.com
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
paths:
...
  /v1/resource-definitions/{resource}/resources:
    parameters:
      - $ref: '#/components/parameters/Resource'
    get:
      tags:
        - developers
      summary: lists all resources for a resource-definition
      operationId: resourceList
      description: |
        Lists all resources for a resource-definition.

        A single resource can be queried by adding its name
        to the resource string like:

        /v1/resource-definitions/rsc1/resources/nodeA
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Array of all resources for the resource-definition
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resource'
    post:
      tags:
        - developers
      summary: create one or more resources.
      operationId: resourceCreate
      description: |
        Adds one or more resource(s).

      responses:
        '201':
          description: resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '500':
          $ref: '#/components/responses/OperationFailed'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ResourceCreate'
            example:
              - resource:
                  name: rsc1
                  node_name: alpha
    delete:
      tags:
        - developers
      summary: truncate a resource-definition
      operationId: resourceDefinitionTruncate
      description: |
        Deletes all resources of the resource-definition ("truncate") without deleting the
        resource-definition itself or any of its snapshots.

        If `delete_empty_resource_definition` is set, the resource-definition is deleted as
        well when it has neither resources nor snapshots left after the truncate. The check
        and the resource-definition deletion are performed atomically, so a concurrently
        created resource or snapshot keeps the resource-definition.
      parameters:
        - in: query
          name: delete_empty_resource_definition
          description: |
            If true, also delete the resource-definition when it has neither resources nor
            snapshots left after its resources have been deleted. The check and the
            resource-definition deletion are performed atomically.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: resources deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
        '500':
          $ref: '#/components/responses/OperationFailed'
  /v1/resource-definitions/{resource}/resources/{node}:
    parameters:
      - $ref: '#/components/parameters/Resource'
      - $ref: '#/components/parameters/Node'
    post:
      tags:
        - developers
      summary: create a resource on a node
      operationId: resourceCreateOnNode
      description: |
        Adds a resource on a node.

        To use a specific storage pool add the `StorPoolName` property
        and use the storage pool name as value.
        If the `StorPoolName` property is not set, the `DfltStorPool` will be used.

        To create a diskless resource you have to set the "DISKLESS" flag in the flags list.
        ```
        {
          "resource": {
            "flags": ["DISKLESS"]
          }
        }
        ```
      responses:
        '201':
          description: resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '500':
          $ref: '#/components/responses/OperationFailed'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreate'
            example: {}
    put:
      tags:
        - developers
      summary: modify a resource
      operationId: resourceModify
      description: |+
        Sets or modifies properties

        Possible properties are:
        - `StorPoolName` - regex[`^[a-zA-Z0-9_][a-zA-Z0-9_-]{2,47}$`]

            Linstor storage pool name to use.

        - `StorPoolNameDrbdMeta` - regex[`^|.internal|[a-zA-Z0-9_][a-zA-Z0-9_-]{2,47}$`]

            Linstor storage pool name to use for external metadata.

        - `PrefNic` - regex[`^[a-zA-Z0-9_][a-zA-Z0-9_-]{2,47}$`]

            Preferred network interface to use

        - `PeerSlots` - range[`1-31`]

            DRBD peer slots present on resource

        - `FileSystem/Type` - enum

            File system type to use

            * ext4
            * xfs

        - `FileSystem/MkfsParams` - string

            Additional parameters for the mkfs command

        - `NVMe/PrefNic` - regex[`^[a-zA-Z0-9_][a-zA-Z0-9_-]{2,47}$`]

            Preferred network interface to use

        - `sys/fs/blkio_throttle_read` - long

            Sets the /sys/fs/cgroup/blkio/blkio.throttle.read_bps_device

        - `sys/fs/blkio_throttle_write` - long

            Sets the /sys/fs/cgroup/blkio/blkio.throttle.write_bps_device

        - `sys/fs/blkio_throttle_read_iops` - long

            Sets the /sys/fs/cgroup/blkio/blkio.throttle.read_iops_device

        - `sys/fs/blkio_throttle_write_iops` - long

            Sets the /sys/fs/cgroup/blkio/blkio.throttle.write_iops_device

        - `DrbdOptions/auto-diskful-allow-cleanup` - boolean_true_false

            Allows this resource to be cleaned up after toggle-disk + resync is finished

      responses:
        '200':
          description: resource modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceModify'
            example:
              override_props:
                Aux/testprop: '1'
              delete_props:
                - Aux/deleteprop
    delete:
      tags:
        - developers
      parameters:
        - name: keep_tiebreaker
          in: query
          description: Ensures that the tiebreaker resource is kept, possibly not deleting the given resource
          schema:
            type: boolean
            default: false
      summary: delete a resource
      operationId: resourceDelete
      description: Delete a resource on a node
      responses:
        '200':
          description: resource deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
  /v1/resource-definitions/{resource}/resources/{node}/make-available:
    parameters:
      - $ref: '#/components/parameters/Resource'
      - $ref: '#/components/parameters/Node'
    post:
      tags:
        - developers
      summary: creates the resource if not already deployed
      operationId: resourceMakeAvailableOnNode
      description: |
        Adds a resource on a node if not already deployed.

        New resources will be diskless, if possible. To force creation of
        a diskful resource, set `diskful` to `true` in the request body.

        Optionally, you can override the layer stack for the created resource.
        If not given, the default configuration for the resource definition is
        used.

        The storage pool will be selected by the autoplacer.
      responses:
        '200':
          description: resource is either created or ensured to be available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
        '500':
          $ref: '#/components/responses/OperationFailed'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceMakeAvailable'
  /v1/resource-definitions/{resource}/resources/{node}/unmake-available:
    parameters:
      - $ref: '#/components/parameters/Resource'
      - $ref: '#/components/parameters/Node'
    post:
      tags:
        - developers
      summary: reverts a make-available issued for a live migration
      operationId: resourceUnmakeAvailableOnNode
      description: |
        Reverts the changes of a make-available with `auto_manage_dual_primary`
        and removes the resource from the given node (the migration source) if
        that can be done without losing data: the resource is deleted if it is
        diskless or if it is a redundant copy in a shared storage pool.
        Tiebreaker and diskful resources are kept.

        Calling this for a resource that is not deployed on the node is a
        successful no-op, so automation clients can always issue this call
        after a migration.
      responses:
        '200':
          description: resource is no longer available on the node
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
        '500':
          $ref: '#/components/responses/OperationFailed'
  /v1/resource-definitions/{resource}/resources/{node}/activate:
    parameters:
      - $ref: '#/components/parameters/Resource'
      - $ref: '#/components/parameters/Node'
    post:
      tags:
        - developers
      summary: activates the resource
      operationId: ActivateRsc
      description: Activates the given resource if possible
      responses:
        '200':
          description: Resource is now active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
        '500':
          $ref: '#/components/responses/OperationFailed'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Resource:
      name: resource
      in: path
      required: true
      description: resource to use
      schema:
        type: string
...
    Offset:
      name: offset
      in: query
      description: number of records to skip for pagination
      schema:
        type: integer
        format: int32
        minimum: 0
    Limit:
      name: limit
      in: query
      description: maximum number of records to return
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 50
...
  responses:
    InvalidInput:
      description: invalid input, or request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiCallRcList'
...
    OperationFailed:
      description: operation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiCallRc'
  schemas:
...
    ApiCallRcList:
      type: array
      items:
        $ref: '#/components/schemas/ApiCallRc'
...